Fix warnings

+++ src/models/extra_inputs.cpp
@@ -44,7 +44,7 @@ ExtraInputs::ExtraInputs(const Model& model, State& state) : model_{model}, stat
 
 void ExtraInputs::Add() {
   // Add extra user inputs to the state
-  for (int i = 0, lim = extra_input_names_.size(); i < lim; ++i) {
+  for (size_t i = 0, lim = extra_input_names_.size(); i < lim; ++i) {
     state_.input_names_.push_back(extra_input_names_[i].c_str());
     state_.inputs_.push_back(extra_inputs_[i].get());
   }
